refactor: 팀 카드 내부 컴포넌트 분리 및 추상화 수준 통일
